Example 66: Can remove elements
So there is something here that I have not yet fully internalized.
The update step needs to be chained off the add step, but the exit
step needs to be chained off the original selection. It has something
to do with the way that add and update interact, but remove doesn't.