Examples 70-73 Iris data scatter plot
70-73 were not really new information. 73 is a basic scatter plot
with data from a csv. I think my biggest take away was the underlying
javascript may not be as immutable as the clojurescript. I was having
a problem where the x domain was being mapped incorrectly in the chart.
I eventually resolved the problem by not reusing the linear scale for
both the x and y axis.