Example 93: Specify people per pixel
Skipped a few exmaples that I had either already implemented the
change or ideomatic D3 was a but more combersome than I wanted to
get into for this practice. The take away from this is we have
specified how many people each pixel should represent and draw
the circle based on that. Also not that the D3 objects are mutable
and that is why we can get the current domain from scale