Example 99: Line chart of time v temp
There was a problem with the domain on the time axis. It was some
interaction between clojurescript and javascript. I was using
js/Date. as the function passed to update, but was getting the
String date back rather than a Date. Switched to an anonymous
function #(js/Date. %1) and got the Date value in the domain.